+++ sdk/azidentity/live_test.go
@@ -6,7 +6,6 @@ package azidentity
 import (
 	"context"
 	"fmt"
-	"log"
 	"net/http"
 	"net/url"
 	"os"
@@ -87,29 +86,29 @@ func init() {
 }
 
 func TestMain(m *testing.M) {
-	switch recording.GetRecordMode() {
-	case recording.PlaybackMode:
-		// enable BodilessMatcher because we don't record request bodies
-		// TODO: add an API for this to sdk/internal
-		req, err := http.NewRequest("POST", "http://localhost:5000/Admin/SetMatcher", http.NoBody)
-		if err != nil {
-			panic(err)
-		}
-		req.Header["x-abstraction-identifier"] = []string{"BodilessMatcher"}
-		res, err := http.DefaultClient.Do(req)
-		if err != nil {
-			panic(err)
-		}
-		if res.StatusCode != http.StatusOK {
-			log.Panicf("failed to enable BodilessMatcher: %v", res)
-		}
-		// TODO: reset matcher
-	case recording.RecordingMode:
-		// remove default sanitizers such as the OAuth response sanitizer
+	if recording.GetRecordMode() == recording.PlaybackMode || recording.GetRecordMode() == recording.RecordingMode {
+		// Start from a fresh proxy
 		err := recording.ResetProxy(nil)
 		if err != nil {
 			panic(err)
 		}
+
+		// At the end of testing we want to reset as to not interfere with other tests.
+		defer func() {
+			err := recording.ResetProxy(nil)
+			if err != nil {
+				panic(err)
+			}
+		}()
+	}
+
+	switch recording.GetRecordMode() {
+	case recording.PlaybackMode:
+		err := recording.SetBodilessMatcher(nil, nil)
+		if err != nil {
+			panic(err)
+		}
+	case recording.RecordingMode:
 		// replace path variables with fake values to simplify matching (the real values aren't secret)
 		pathVars := map[string]string{
 			liveManagedIdentity.clientID:                    fakeClientID,
@@ -120,7 +119,7 @@ func TestMain(m *testing.M) {
 		}
 		for target, replacement := range pathVars {
 			if target != "" {
-				err = recording.AddURISanitizer(replacement, target, nil)
+				err := recording.AddURISanitizer(replacement, target, nil)
 				if err != nil {
 					panic(err)
 				}
@@ -139,7 +138,7 @@ func TestMain(m *testing.M) {
 		// remove token request bodies (which are form encoded) because they contain
 		// secrets, are irrelevant in matching, and are formed by MSAL anyway
 		// (note: Cloud Shell would need an exemption from this, and that would be okay--its requests contain no secrets)
-		err = recording.AddBodyRegexSanitizer("{}", `^\S+=.*`, nil)
+		err := recording.AddBodyRegexSanitizer("{}", `^\S+=.*`, nil)
 		if err != nil {
 			panic(err)
 		}
@@ -150,15 +149,10 @@ func TestMain(m *testing.M) {
 				panic(err)
 			}
 		}
-		defer func() {
-			err := recording.ResetProxy(nil)
-			if err != nil {
-				panic(err)
-			}
-			// TODO: reset matcher
-		}()
 	}
-	os.Exit(m.Run())
+	val := m.Run()
+
+	os.Exit(val)
 }
 
 func initRecording(t *testing.T) (policy.ClientOptions, func()) {