+++ sdk/keyvault/azsecrets/internal/constants.go
@@ -10,7 +10,7 @@ package internal
 
 const (
 	moduleName    = "internal"
-	moduleVersion = "v0.5.0"
+	moduleVersion = "v0.5.1"
 )
 
 // DataAction - Supported permissions for data actions.