+++ sdk/messaging/azservicebus/client_test.go
@@ -5,6 +5,7 @@ package azservicebus
 
 import (
 	"context"
+	"fmt"
 	"net"
 	"os"
 	"testing"
@@ -23,13 +24,29 @@ func TestNewClientWithAzureIdentity(t *testing.T) {
 
 	// test with azure identity support
 	ns := os.Getenv("SERVICEBUS_ENDPOINT")
+
+	var credsToAdd []azcore.TokenCredential
+
+	cliCred, err := azidentity.NewAzureCLICredential(nil)
+	require.NoError(t, err)
+
 	envCred, err := azidentity.NewEnvironmentCredential(nil)
+
+	if err == nil {
+		fmt.Printf("Env cred works, being added to our chained token credential")
+		credsToAdd = append(credsToAdd, envCred)
+	}
+
+	credsToAdd = append(credsToAdd, cliCred)
+
+	cred, err := azidentity.NewChainedTokenCredential(credsToAdd, nil)
+	require.NoError(t, err)
 
 	if err != nil || ns == "" {
 		t.Skip("Azure Identity compatible credentials not configured")
 	}
 
-	client, err := NewClient(ns, envCred, nil)
+	client, err := NewClient(ns, cred, nil)
 	require.NoError(t, err)
 
 	sender, err := client.NewSender(queue, nil)