+++ sdk/messaging/azservicebus/liveTestHelpers_test.go
@@ -5,12 +5,14 @@ package azservicebus
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"testing"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus/admin"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus/internal/test"
+	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus/internal/utils"
 	"github.com/stretchr/testify/require"
 )
 
@@ -64,4 +66,30 @@ func deleteQueue(t *testing.T, ac *admin.Client, queueName string) {
 func deleteSubscription(t *testing.T, ac *admin.Client, topicName string, subscriptionName string) {
 	_, err := ac.DeleteSubscription(context.Background(), topicName, subscriptionName, nil)
 	require.NoError(t, err)
+}
+
+// peekSingleMessageForTest wraps a standard Receiver.Peek() call so it returns at least one message
+// and fails tests otherwise.
+func peekSingleMessageForTest(t *testing.T, receiver *Receiver) *ReceivedMessage {
+	var msg *ReceivedMessage
+
+	// Peek, unlike Receive, doesn't block until at least one message has arrived, so we have to poll
+	// to get a similar effect.
+	err := utils.Retry(context.Background(), "peek", func(ctx context.Context, args *utils.RetryFnArgs) error {
+		peekedMessages, err := receiver.PeekMessages(context.Background(), 1, nil)
+		require.NoError(t, err)
+
+		if len(peekedMessages) == 1 {
+			msg = peekedMessages[0]
+			return nil
+		} else {
+			return errors.New("No peekable messages available")
+		}
+	}, func(err error) bool {
+		return false
+	}, utils.RetryOptions{})
+
+	require.NoError(t, err)
+
+	return msg
 }