+++ sdk/internal/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.9.1 (Unreleased)
+## 0.9.2 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,11 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 0.9.1 (2022-02-01)
+
+### Features Added
+* Adds a `CustomDefaultMatcher` that adds headers `:path`, `:authority`, `:method`, and `:scheme` to the default matcher.
 
 ## 0.9.0 (2022-01-24)