+++ sdk/keyvault/azsecrets/CHANGELOG.md
@@ -1,15 +1,19 @@
 # Release History
 
-## 0.5.0 (Unreleased)
+## 0.5.1 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
-* Fixes a bug where `UpdateSecretProperties` will delete properties that are not explicitly set each time. This is only a breaking change at runtime, where the request body will change.
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 0.5.0 (2022-02-08)
+
+### Breaking Changes
+* Fixes a bug where `UpdateSecretProperties` will delete properties that are not explicitly set each time. This is only a breaking change at runtime, where the request body will change.
 
 ## 0.4.0 (2022-01-11)